requirements.txt
Django==3.1.2
ipdb
requests
petl
python-dateutil==2.8.1